Quality control

+++ db/migrate/20170906081925_create_categories.rb
@@ -2,8 +2,8 @@ class CreateCategories < ActiveRecord::Migration[5.1]
   def change
     create_table :categories do |t|
       t.string :name
-      t.decimal :place
-      t.decimal :runs
+      t.integer :place
+      t.integer :runs
 
       t.timestamps
     end
