+++ db/migrate/20170906081945_create_results.rb
@@ -2,7 +2,7 @@ class CreateResults < ActiveRecord::Migration[5.1]
   def change
     create_table :results do |t|
       t.decimal :score
-      t.decimal :run
+      t.integer :run
       t.references :parameter, foreign_key: true
       t.references :quality_control, foreign_key: true
